Fix transmitting conditions
Needlessly thorough fix for a bad index problem in one-block
transmitting conditions.  Also cleans up EMF stuff back into
B_CT package, and sets EMF averaging to default with
transmitting conditions now it's decently fast.
