+++ kharma/boundaries/one_block_transmit.cpp
@@ -62,13 +62,20 @@ void KBoundaries::TransmitSetTE(MeshBlockData<Real> *rc, VariablePack<Real> &q,
 {
     // We're sometimes called without any variables to sync (e.g. syncing flags, EMFs), just return
     if (q.GetDim(4) == 0) return;
+    // We're also sometimes called on coarse buffers with or without AMR.
+    // Use of transmitting polar conditions when coarse buffers matter (e.g., refinement
+    // boundary touching the pole) is UNSUPPORTED
+    if (coarse) return;
 
-    // Indices
+    // Pull boundary properties
     auto pmb = rc->GetBlockPointer();
     const bool binner = BoundaryIsInner(bface);
-    const int dir = BoundaryDirection(bface);
+    const int bdir = BoundaryDirection(bface);
     const auto domain = BoundaryDomain(bface);
     const auto bname = BoundaryName(bface);
+
+    if (bdir != 2)
+        throw std::runtime_error("Transmitting polar conditions only defined for X2!");
 
     std::vector<TopologicalElement> el_list;
     if (do_face) {
@@ -76,41 +83,37 @@ void KBoundaries::TransmitSetTE(MeshBlockData<Real> *rc, VariablePack<Real> &q,
     } else {
         el_list = {CC};
     }
-    int el_tot = el_list.size();
-    for (auto el : el_list) {
-        // Set boundary/ghost zones *only*, not zones on faces
-        const IndexRange3 b = KDomain::GetRange(rc, domain, el, coarse);
-        const IndexRange3 bi = KDomain::GetRange(rc, IndexDomain::interior, CC);
+    for (TopologicalElement &el : el_list) {
+        // This automatically includes zones on domain faces, which we set to 0 below
+        const IndexRange3 b = KDomain::GetBoundaryRange(rc, domain, el, coarse);
+        const IndexRange3 bi = KDomain::GetRange(rc, IndexDomain::interior, CC, coarse);
+        // Whether we're on e.g. X2 face for X2 boundary
+        const bool corresponding_face = (el == FaceOf(bdir));
+        const int reflect_offset = corresponding_face ? 0 : (binner ? 1 : -1);
 
-        if (domain == IndexDomain::inner_x2 || domain == IndexDomain::outer_x2) {
-            const int Nk3p = (bi.ke - bi.ks + 1); // Physical/interior *zones* in dir 3
-            const int Nk3p2 = Nk3p/2;             // pi/2 of those (boundary incompatible with slice sims TODO check+error)
-            const int ksp = bi.ks;                // Offset of first physical zone or face (same number)
-            // Pivot element for faces is first domain face (==0), pivot for cells is between b.js/e, b.js/e+/-1
-            const int jpivot = (domain == IndexDomain::inner_x2) ? ((el == FaceOf(dir)) ? b.je + 1 : b.je)
-                                                                 : ((el == FaceOf(dir)) ? b.js - 1 : b.js);
-            const bool do_face_invert = (el == F3);
-            pmb->par_for(
-                "transmitting_polar_boundary_" + bname, 0, q.GetDim(4)/el_tot-1, b.ks, b.ke, b.js, b.je, b.is, b.ie,
-                KOKKOS_LAMBDA (const int &v, const int &k, const int &j, const int &i) {
-                    const int ki = ((k - ksp + Nk3p2) % Nk3p) + ksp;
-                    const int ji = m::abs(jpivot - j);
-                    const int ii = i;
-                    const Real invert = (do_face_invert || q(el, v).vector_component == X3DIR) ? -1. : 1.;
-                    q(el, v, k, j, i) = invert * q(el, v, ki, ji, ii);
-                }
-            );
-            // Explicitly zero B2 face for some reason
-            if (el == FaceOf(dir)) {
-                pmb->par_for(
-                    "transmitting_polar_boundary_" + bname, 0, q.GetDim(4)/el_tot-1, b.ks, b.ke, jpivot, jpivot, b.is, b.ie,
-                    KOKKOS_LAMBDA (const int &v, const int &k, const int &j, const int &i) {
-                        q(el, v, k, j, i) = 0.;
-                    }
-                );
+        // TODO SPECIFIC TO X2
+        // Total physical/interior *zones* in dir 3.  Note first/last face will share an opposite,
+        // i.e. both F3(0,x,x) and F3(N3,x,x) -> F3(N3/2,x,x), but in reverse only the former
+        const int Nk3p = (bi.ke - bi.ks + 1);
+        const int Nk3p2 = Nk3p/2;
+        const int ksp = bi.ks;
+        // Pivot element for corresponding face is one more than pivot for cells/other-faces
+        // X2 dir should be calculated exactly as if reflecting
+        const int jpivot = (binner) ? (corresponding_face ? b.je + 1 : b.je)
+                                    : (corresponding_face ? b.js - 1 : b.js);
+        // B3 component on X3 face should be inverted even if not marked "vector"
+        // TODO currently this inverts everything on F3, probably not desirable
+        const bool do_face_invert = (el == F3);
+
+        pmb->par_for(
+            "transmitting_polar_boundary_" + bname, 0, q.GetDim(4)-1, b.ks, b.ke, b.js, b.je, b.is, b.ie,
+            KOKKOS_LAMBDA (const int &v, const int &k, const int &j, const int &i) {
+                const int ki = ((k - ksp + Nk3p2) % Nk3p) + ksp;
+                const int ji = jpivot + reflect_offset + (jpivot - j);
+                const int ii = i;
+                const Real invert = (do_face_invert || q(el, v).vector_component == X3DIR) ? -1. : 1.;
+                q(el, v, k, j, i) = (corresponding_face && j == jpivot) ? 0. : invert * q(el, v, ki, ji, ii);
             }
-        } else {
-            throw std::runtime_error("Transmitting polar conditions only defined for X2!");
-        }
+        );
     }
 }
